+++ src/app/(pages)/queryBuilding/buildFromTemplates/BuildFromTemplates.tsx
@@ -40,12 +40,12 @@ import {
   getConditionsData,
   getCustomQueries,
   getSavedQueryById,
+  getValueSetsAndConceptsByConditionIDs,
   saveCustomQuery,
 } from "@/app/backend/query-building/service";
 import { useSession } from "next-auth/react";
 import { ModalRef } from "@/app/ui/designSystem/modal/Modal";
 import WarningModal from "@/app/ui/designSystem/modal/warningModal";
-import { getValueSetsAndConceptsByConditionIDs } from "@/app/backend/db-creation/service";
 
 export type FormError = {
   queryName: boolean;
