+++ src/app/backend/query-building/service.ts
@@ -20,6 +20,7 @@ import { CustomUserQuery } from "@/app/models/entities/query";
 import { DibbsValueSet } from "@/app/models/entities/valuesets";
 import { auditable } from "../audit-logs/decorator";
 import { linkTimeboxRangesToQuery } from "../query-timefiltering";
+import { getValueSetsByConditionIdsSql } from "../db-creation/seedSqlStructs";
 
 class QueryBuildingService {
   /**
@@ -237,6 +238,37 @@ class QueryBuildingService {
   }
 
   /**
+   * Executes a search for a ValueSets and Concepts against the Postgres
+   * Database, using the ID of the condition associated with any such data.
+   * @param ids Array of ids for entries in the conditions table
+   * @returns One or more rows from the DB matching the requested saved query,
+   * or an error if no results can be found.
+   */
+  static async getValueSetsAndConceptsByConditionIDs(ids: string[]) {
+    try {
+      if (ids.length === 0) {
+        throw Error("No condition ids passed in to query by");
+      }
+
+      const escapedValues = ids.map((_, i) => `$${i + 1}`).join() + ")";
+      const queryString = getValueSetsByConditionIdsSql + escapedValues;
+
+      const result = await dbService.query(queryString, ids);
+      if (result.rows.length === 0) {
+        console.error("No results found for given condition ids", ids);
+        return [];
+      }
+      return result.rows;
+    } catch (error) {
+      console.error(
+        "Error retrieving value sets and concepts for condition",
+        error,
+      );
+      throw error;
+    }
+  }
+
+  /**
    * Retrieves all records from the conditions table in the database.
    * This function queries the database to fetch condition data, including
    * condition name, code, and category.
@@ -285,3 +317,5 @@ export const getQueryList = QueryBuildingService.getQueryList;
 export const getQueryById = QueryBuildingService.getQueryById;
 export const getQueriesForUser = QueryBuildingService.getQueriesForUser;
 export const getConditionsData = QueryBuildingService.getConditionsData;
+export const getValueSetsAndConceptsByConditionIDs =
+  QueryBuildingService.getValueSetsAndConceptsByConditionIDs;
